Cascade animation for layers; Fixed delete video
Fixed delete video: When the selected is CED or CID, the video is removed. When it is CBV or CEV, it is added. Previously the "element amount" wasn't updated. And when choosing video format after a display format, it didn't add the video. I also added display:none to the video's delete button.